Add modifications to receive position in double
Modified the a C test function to test correctly with the new double getPOSI

+++ C/monitorExample/main.c
@@ -55,7 +55,6 @@ struct timeval tv;
 
 int waitForInput()
 {
-
 	FD_ZERO(&fds);
 	FD_SET(fdstdin, &fds);
 	select(1, &fds, NULL, NULL, &tv);
@@ -70,7 +69,7 @@ int main(void)
     tv.tv_usec = 100 * 1000;
 	while (1)
 	{
-		float posi[7]; // FIXME: change this to the 64-bit lat/lon/h
+		double posi[7]; // FIXME: change this to the 64-bit lat/lon/h
 		int result = getPOSI(client, posi, aircraftNum);
 		if (result < 0) // Error in getPOSI
 		{
@@ -84,7 +83,7 @@ int main(void)
 			break;
 		}
 
-		printf("Loc: (%4f, %4f, %4f) Aileron:%2f Elevator:%2f Rudder:%2f\n",
+		printf("Loc: (%4lf, %4lf, %4lf) Aileron:%2f Elevator:%2f Rudder:%2f\n",
 			posi[0], posi[1], posi[2], ctrl[1], ctrl[0], ctrl[2]);
 
 		// Check if any key has been pressed and break
